Add unicorn to Gemfile.
The generated files did not include the unicorn
gem. When navigating (host) to localhost:8080 a
"Something went wrong" error is displayed.

To check what went wrong run (guest) the command:

$ sudo tail -f /var/log/upstart/railstutorial.log

To fix the problem the unicorn gem is added and
run (guest):

/railstutorial$ bundler

Now navigating (host) to localhost:8080 shows the
default root page.